Move modal add location & add summoner to containers
